Fix-up i686/armv7 support

+++ test/runtests.jl
@@ -120,11 +120,11 @@ end
 end
 
 @testset "_addcommas" begin
-    @test LinuxPerf._addcommas(1) == "1"
-    @test LinuxPerf._addcommas(12) == "12"
-    @test LinuxPerf._addcommas(123) == "123"
-    @test LinuxPerf._addcommas(1234) == "1,234"
-    @test LinuxPerf._addcommas(12345) == "12,345"
+    @test LinuxPerf._addcommas(Int64(1)) == "1"
+    @test LinuxPerf._addcommas(Int64(12)) == "12"
+    @test LinuxPerf._addcommas(Int64(123)) == "123"
+    @test LinuxPerf._addcommas(Int64(1234)) == "1,234"
+    @test LinuxPerf._addcommas(Int64(12345)) == "12,345"
     @test LinuxPerf._addcommas(typemin(Int64)) == "-9,223,372,036,854,775,808"
 end
